refactor: grouping rule related commands
