+++ lib/mihari/commands/rule.rb
@@ -4,20 +4,41 @@ require "pathname"
 
 module Mihari
   module Commands
-    module Initializer
+    module Rule
       def self.included(thor)
         thor.class_eval do
-          desc "init", "Initialize a new rule"
-          method_option :path, type: :string, default: "./rule.yml"
-          def init
-            path = options["path"]
+          desc "validate [PATH]", "Validate a rule file"
+          #
+          # Validate format of a rule
+          #
+          # @param [String] path
+          #
+          def validate(path)
+            rule = Structs::Rule.from_path_or_id(path)
 
+            begin
+              rule.validate!
+              Mihari.logger.info "Valid format. The input is parsed as the following:"
+              Mihari.logger.info rule.data.to_yaml
+            rescue RuleValidationError
+              nil
+            end
+          end
+
+          desc "init [PATH]", "Initialize a new rule file"
+          #
+          # Initialize a new rule file
+          #
+          # @param [String] path
+          #
+          #
+          def init(path = "./rule.yml")
             warning = "#{path} exists. Do you want to overwrite it? (y/n)"
             return if Pathname(path).exist? && !(yes? warning)
 
             initialize_rule path
 
-            Mihari.logger.info "A new rule is initialized as #{path}."
+            Mihari.logger.info "A new rule is initialized: #{path}."
           end
 
           no_commands do
